manage restaurant APIs

+++ Laravel-back-end/database/migrations/restaurant/2023_08_31_2_create_tables_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->foreignId("branch_id")->constrained()->cascadeOnDelete();
-            $table->tinyInteger("capacity");
+            $table->integer("capacity");
             $table->timestamps();
         });
     }
